forecast command: ApplicationCommand-only
